Draw the grid

+++ Assets/Scripts/Influence.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Influence : MonoBehaviour {
 
-    public float strength; // black 4, yellow 3, blue 2, white 1
+    public int strength; // black 4, yellow 3, blue 2, white 1
     public bool side; // 1 is red side, 0 is green
 
 	// Use this for initialization
